Refactor GsmClientBG96 and GsmClientSecureBG96 constructors
This commit refactors the constructors of the GsmClientBG96 and GsmClientSecureBG96 classes in the TinyGsmClientBG96.h file. The changes initialize the ssl_sock variable to false in the GsmClientBG96 constructor and to true in the GsmClientSecureBG96 constructor. This ensures that the correct socket type is set when creating instances of these classes.